Split `Rollout` into a separate module
